+++ Test/Integration/Controller/Adminhtml/Ajax/CustomerSearchTest.php
@@ -16,10 +16,10 @@ class CustomerSearchTest extends AbstractAjaxTestCase
      * Setup method
      * @throws AuthenticationException
      */
-    public function setUp()
+    protected function setUp(): void
     {
-        $this->uri = 'backend/emailtester/ajax/customerSearch';
         parent::setUp();
+        $this->uri = 'backend/emailtester/ajax/customerSearch';
     }
 
     /**