+++ Test/Integration/Controller/Adminhtml/Ajax/OrderLabelTest.php
@@ -14,10 +14,10 @@ class OrderLabelTest extends AbstractAjaxTestCase
      * Setup method
      * @throws AuthenticationException
      */
-    public function setUp()
+    protected function setUp(): void
     {
-        $this->uri = 'backend/emailtester/ajax/orderLabel';
         parent::setUp();
+        $this->uri = 'backend/emailtester/ajax/orderLabel';
     }
 
     /**