+++ Test/Integration/Controller/Adminhtml/Ajax/ProductSearchTest.php
@@ -14,10 +14,10 @@ class ProductSearchTest extends AbstractAjaxTestCase
      * Setup method
      * @throws AuthenticationException
      */
-    public function setUp()
+    protected function setUp(): void
     {
-        $this->uri = 'backend/emailtester/ajax/productSearch';
         parent::setUp();
+        $this->uri = 'backend/emailtester/ajax/productSearch';
     }
 
     /**