+++ Test/Unit/Behaviour/ErrorableTest.php
@@ -24,8 +24,9 @@ class ErrorableTest extends TestCase
     /**
      * Setup method
      */
-    protected function setUp()
+    protected function setUp(): void
     {
+        parent::setUp();
         $this->targetTrait = $this->getMockForTrait(Target::class);
     }