+++ Test/Unit/Helper/OutputTest.php
@@ -36,8 +36,10 @@ class OutputTest extends TestCase
     /**
      * Setup method
      */
-    protected function setUp()
+    protected function setUp(): void
     {
+        parent::setUp();
+
         $this->objectManagerHelper = new ObjectManagerHelper($this);
 
         $context = $this->getContextStub();